Changes to LoginScreen, Adjusted database package
