+++ src/main/java/HelloWorld.java
@@ -1,5 +1,3 @@
-package tutorial;
-
 public class HelloWorld {
 
     public static void main(String[] args) {
