+++ src/main/java/database/EventFileUser.java
@@ -1,60 +1,28 @@
-package tutorial;
+package database;
 
 import java.sql.*;
 import java.util.ArrayList;
 
-public class ParFileUser {
-    public static void main(String[] args) {
-    }
+public class EventFileUser {
 
-    public void utilStorePar(String username, String password){
+    public void utilStoreEvent(String title,
+                               int status,
+                               int event_type,
+                               String description,
+                               String location,
+                               String image_path,
+                               int year,
+                               int month,
+                               int day,
+                               int hour,
+                               int minute){
         Statement stmt = null;
         Connection conn = null;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "insert into parfile(username, password) values('" + username + "','" + password + "');" ;
-            stmt = conn.createStatement();
-            int count = stmt.executeUpdate(sql);
-            System.out.println(sql);
-            if (count > 0 ){
-                System.out.println("Success");
-            }
-            else {
-                System.out.println("Failure");
-            }
-
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            if (stmt != null){
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null){
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-
-    }
-
-    public void utilDeletePar(String username){
-        Statement stmt = null;
-        Connection conn = null;
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "delete from parfile where username = '" + username + "';";
+            String sql = "insert into eventfile(title,status,event_type,description,location,image_path,year,month,day,hour,minute) values('" +
+                    title + "'," + status + "," + event_type + ",'" + description + "','" + location + "','" + image_path + "'," + year + "," + month + "," + day + "," + hour + "," + minute + ");";
             stmt = conn.createStatement();
             int count = stmt.executeUpdate(sql);
             System.out.println(sql);
@@ -87,14 +55,13 @@ public class ParFileUser {
 
         }
     }
-
-    public void utiladdparfollowing(String par_username, String org_username){
+    public void utilDeleteEvent(String title){
         Statement stmt = null;
         Connection conn = null;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "insert into follow_org_par(par_username, org_username) values('" + par_username + "','" + org_username + "');" ;
+            String sql = "delete from eventfile where title = '" + title + "';";
             stmt = conn.createStatement();
             int count = stmt.executeUpdate(sql);
             System.out.println(sql);
@@ -126,65 +93,36 @@ public class ParFileUser {
             }
 
         }
-
-
     }
 
-    public void utilDeleteParFollowOrg(String par_username, String org_username) {
-            Statement stmt = null;
-            Connection conn = null;
-            try {
-                Class.forName("com.mysql.jdbc.Driver");
-                conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-                String sql = "delete from follow_org_par where par_username = '" + par_username + "' and org_username = '" + org_username + "';";
-                stmt = conn.createStatement();
-                int count = stmt.executeUpdate(sql);
-                System.out.println(sql);
-                if (count > 0) {
-                    System.out.println("Success");
-                } else {
-                    System.out.println("Failure");
-                }
-
-            } catch (ClassNotFoundException e) {
-                e.printStackTrace();
-            } catch (SQLException e) {
-                e.printStackTrace();
-            } finally {
-                if (stmt != null) {
-                    try {
-                        stmt.close();
-                    } catch (SQLException e) {
-                        e.printStackTrace();
-                    }
-                }
-                if (conn != null) {
-                    try {
-                        conn.close();
-                    } catch (SQLException e) {
-                        e.printStackTrace();
-                    }
-                }
-
-            }
-
-        }
-
-    public void utilAddParPastevent(String par_username, String event_title){
+    public String utilGetOrganizer(String title){
         Statement stmt = null;
         Connection conn = null;
+        ResultSet rs = null;
+        String organizer = null;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "insert into past_events_for_par(par_username, event_title) values('" + par_username + "','" + event_title + "');" ;
             stmt = conn.createStatement();
-            int count = stmt.executeUpdate(sql);
-            System.out.println(sql);
-            if (count > 0 ){
-                System.out.println("Success");
+            try (ResultSet unpublished_organizer = stmt.executeQuery("select org_username from unpublished_events_for_org where event_title = '" + title + "';")){
+                if (unpublished_organizer.next()){
+                    organizer = unpublished_organizer.getString(1);
+                    unpublished_organizer.close();
+                };
             }
-            else {
-                System.out.println("Failure");
+            try (ResultSet past_organizer = stmt.executeQuery("select org_username from past_events_for_org where event_title = '" + title + "';")){
+                if (past_organizer.next()){
+                    organizer = past_organizer.getString(1);
+                    past_organizer.close();
+
+                };
+            }
+
+            try (ResultSet upcoming_organizer = stmt.executeQuery("select org_username from past_events_for_org where event_title = '" + title + "';")){
+                if (upcoming_organizer.next()){
+                    organizer = upcoming_organizer.getString(1);
+                    upcoming_organizer.close();
+                }
             }
 
         } catch (ClassNotFoundException e) {
@@ -192,6 +130,13 @@ public class ParFileUser {
         } catch (SQLException e) {
             e.printStackTrace();
         }finally {
+            if (rs != null){
+                try {
+                    rs.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
             if (stmt != null){
                 try {
                     stmt.close();
@@ -208,134 +153,15 @@ public class ParFileUser {
             }
 
         }
+        return organizer;
 
-
-    }
-    public void utilDeleteParPastevent(String par_username, String event_title){
-        Statement stmt = null;
-        Connection conn = null;
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "delete from past_events_for_par where par_username = '" + par_username + "' and event_title = '" + event_title + "';";
-            stmt = conn.createStatement();
-            int count = stmt.executeUpdate(sql);
-            System.out.println(sql);
-            if (count > 0) {
-                System.out.println("Success");
-            } else {
-                System.out.println("Failure");
-            }
-
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        } finally {
-            if (stmt != null) {
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null) {
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
 
 
 
     }
 
-    public void utilAddParUpcomingEvent(String par_username, String event_title){
-        Statement stmt = null;
-        Connection conn = null;
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "insert into upcoming_events_for_par(par_username, event_title) values('" + par_username + "','" + event_title + "');" ;
-            stmt = conn.createStatement();
-            int count = stmt.executeUpdate(sql);
-            System.out.println(sql);
-            if (count > 0 ){
-                System.out.println("Success");
-            }
-            else {
-                System.out.println("Failure");
-            }
 
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            if (stmt != null){
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null){
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-
-
-    }
-
-    public void utilDeleteParUpcomingevent(String par_username, String event_title){
-        Statement stmt = null;
-        Connection conn = null;
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "delete from upcoming_events_for_par where par_username = '" + par_username + "' and event_title = '" + event_title + "';";
-            stmt = conn.createStatement();
-            int count = stmt.executeUpdate(sql);
-            System.out.println(sql);
-            if (count > 0) {
-                System.out.println("Success");
-            } else {
-                System.out.println("Failure");
-            }
-
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        } finally {
-            if (stmt != null) {
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null) {
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-
-    }
-
-    public ArrayList<String> utilGetAllFollowing(String par_username){
+    public ArrayList<String> utilGetAllPastEventParticipant(String title){
         Statement stmt = null;
         Connection conn = null;
         ResultSet rs = null;
@@ -343,11 +169,8 @@ public class ParFileUser {
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select org_username from follow_org_par where par_username = '" + par_username + "';";
             stmt = conn.createStatement();
-            rs = stmt.executeQuery(sql);
-            System.out.println(sql);
-//            rs.next();
+            rs = stmt.executeQuery("select par_username from past_events_for_par where event_title = '" + title + "';");
             while (rs.next()){
                 l.add(rs.getString(1));
             }
@@ -382,10 +205,8 @@ public class ParFileUser {
         return l;
 
 
-
     }
-
-    public ArrayList<String> utilGetAllPastEvent(String par_username){
+    public ArrayList<String> utilGetAllUpcomingEventParticipant(String title){
         Statement stmt = null;
         Connection conn = null;
         ResultSet rs = null;
@@ -393,10 +214,8 @@ public class ParFileUser {
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select event_title from past_events_for_par where par_username = '" + par_username + "';";
             stmt = conn.createStatement();
-            rs = stmt.executeQuery(sql);
-            System.out.println(sql);
+            rs = stmt.executeQuery("select par_username from upcoming_events_for_par where event_title = '" + title + "';");
             while (rs.next()){
                 l.add(rs.getString(1));
             }
@@ -430,167 +249,26 @@ public class ParFileUser {
         }
         return l;
 
+
     }
-    public ArrayList<String> utilGetAllUpcomingEvent(String par_username){
+
+
+
+
+
+    public int getStatus(String title){
         Statement stmt = null;
         Connection conn = null;
         ResultSet rs = null;
-        ArrayList l = new ArrayList<String>(0);
+        int status = -1;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select event_title from upcoming_events_for_par where par_username = '" + par_username + "';";
-            stmt = conn.createStatement();
-            rs = stmt.executeQuery(sql);
-            System.out.println(sql);
-            while (rs.next()){
-                l.add(rs.getString(1));
-            }
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            if (rs != null){
-                try {
-                    rs.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (stmt != null){
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null){
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-        return l;
-
-    }
-
-    public void utilPasswordUpdating(String par_username, String new_password){
-        Statement stmt = null;
-        Connection conn = null;
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "update parfile set password = '" + new_password + "' where username = '" + par_username + "';";
-            stmt = conn.createStatement();
-            int count = stmt.executeUpdate(sql);
-            System.out.println(sql);
-            if (count > 0) {
-                System.out.println("Success");
-            } else {
-                System.out.println("Failure");
-            }
-
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        } finally {
-            if (stmt != null) {
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null) {
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-    }
-    public void utilNotificationUpdating(String par_username, String new_notification){
-        Statement stmt = null;
-        Connection conn = null;
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "update parfile set notification = '" + new_notification + "' where username = '" + par_username + "';";
-            stmt = conn.createStatement();
-            int count = stmt.executeUpdate(sql);
-            System.out.println(sql);
-            if (count > 0) {
-                System.out.println("Success");
-            } else {
-                System.out.println("Failure");
-            }
-
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        } finally {
-            if (stmt != null) {
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null) {
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-    }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-    public int getPassword(String username){
-        Statement stmt = null;
-        Connection conn = null;
-        ResultSet rs = null;
-        int password = -1;
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select password from parfile where username = '" + username + "';";
+            String sql = "select status from eventfile where title = '" + title + "';";
             stmt = conn.createStatement();
             rs = stmt.executeQuery(sql);
             rs.next();
-            password = rs.getInt("password");
+            status = rs.getInt("status");
         } catch (ClassNotFoundException e) {
             System.out.println("NotFound");
             e.printStackTrace();
@@ -620,21 +298,23 @@ public class ParFileUser {
                 }
             }
         }
-        return password;
+        return status;
+
     }
-    public String getNotification(String username){
+
+    public int getType(String title){
         Statement stmt = null;
         Connection conn = null;
         ResultSet rs = null;
-        String notification = null;
+        int type = -1;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select notification from parfile where username = '" + username + "';";
+            String sql = "select event_type from eventfile where title = '" + title + "';";
             stmt = conn.createStatement();
             rs = stmt.executeQuery(sql);
             rs.next();
-            notification = rs.getString("notification");
+            type = rs.getInt("event_type");
         } catch (ClassNotFoundException e) {
             System.out.println("NotFound");
             e.printStackTrace();
@@ -664,43 +344,220 @@ public class ParFileUser {
                 }
             }
         }
-        return notification;
+        return type;
+
+
     }
 
-    public ArrayList<String> getUpcomingEvents(String username){
-        return utilGetAllUpcomingEvent(username);
+    public String getDescription(String title){
+        Statement stmt = null;
+        Connection conn = null;
+        ResultSet rs = null;
+        String description = null;
+        try {
+            Class.forName("com.mysql.jdbc.Driver");
+            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "select description from eventfile where title = '" + title + "';";
+            stmt = conn.createStatement();
+            rs = stmt.executeQuery(sql);
+            rs.next();
+            description = rs.getString("description");
+        } catch (ClassNotFoundException e) {
+            System.out.println("NotFound");
+            e.printStackTrace();
+        } catch (SQLException e) {
+            System.out.println("SQL");
+            e.printStackTrace();
+        }finally {
+            if(rs != null){
+                try {
+                    rs.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(stmt != null){
+                try {
+                    stmt.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(conn != null){
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+        return description;
+
+
     }
 
-    public ArrayList<String> getPastEvents(String username){
-        return utilGetAllPastEvent(username);
+    public String getLocation(String title){
+        Statement stmt = null;
+        Connection conn = null;
+        ResultSet rs = null;
+        String location = null;
+        try {
+            Class.forName("com.mysql.jdbc.Driver");
+            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "select description from eventfile where title = '" + title + "';";
+            stmt = conn.createStatement();
+            rs = stmt.executeQuery(sql);
+            rs.next();
+            location = rs.getString("location");
+        } catch (ClassNotFoundException e) {
+            System.out.println("NotFound");
+            e.printStackTrace();
+        } catch (SQLException e) {
+            System.out.println("SQL");
+            e.printStackTrace();
+        }finally {
+            if(rs != null){
+                try {
+                    rs.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(stmt != null){
+                try {
+                    stmt.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(conn != null){
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+        return location;
+
+
+
     }
 
-    public ArrayList<String> getFollowedOrg(String username){
-        return utilGetAllFollowing(username);
+    public String getImagePath(String title){
+        Statement stmt = null;
+        Connection conn = null;
+        ResultSet rs = null;
+        String image_path = null;
+        try {
+            Class.forName("com.mysql.jdbc.Driver");
+            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "select image_path from eventfile where title = '" + title + "';";
+            stmt = conn.createStatement();
+            rs = stmt.executeQuery(sql);
+            rs.next();
+            image_path = rs.getString("image_path");
+        } catch (ClassNotFoundException e) {
+            System.out.println("NotFound");
+            e.printStackTrace();
+        } catch (SQLException e) {
+            System.out.println("SQL");
+            e.printStackTrace();
+        }finally {
+            if(rs != null){
+                try {
+                    rs.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(stmt != null){
+                try {
+                    stmt.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(conn != null){
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+        return image_path;
     }
 
-    public void setPassword(String username, String new_password){
-        utilPasswordUpdating(username, new_password);
+    public ArrayList<Integer> getTime(String title){
+        Statement stmt = null;
+        Connection conn = null;
+        ResultSet rs = null;
+        ArrayList<Integer> l = new ArrayList<Integer>(0);
+        try {
+            Class.forName("com.mysql.jdbc.Driver");
+            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "select year,month,day,hour,minute from eventfile where title = '" + title + "';";
+            stmt = conn.createStatement();
+            rs = stmt.executeQuery(sql);
+            rs.next();
+            l.add(rs.getInt("year"));
+            l.add(rs.getInt("month"));
+            l.add(rs.getInt("day"));
+            l.add(rs.getInt("hour"));
+            l.add(rs.getInt("minute"));
+        } catch (ClassNotFoundException e) {
+            System.out.println("NotFound");
+            e.printStackTrace();
+        } catch (SQLException e) {
+            System.out.println("SQL");
+            e.printStackTrace();
+        }finally {
+            if(rs != null){
+                try {
+                    rs.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(stmt != null){
+                try {
+                    stmt.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if(conn != null){
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+        return l;
     }
 
-    public void setNofication(String username, String new_notification){
-        utilNotificationUpdating(username,new_notification);
-    }
-
-    public void followOrg(String par_username, String org_username){
-        utiladdparfollowing(par_username,org_username);
-    }
-
-    public void registerEvent(String par_username, String title){
-        utilAddParUpcomingEvent(par_username,title);
-    }
-
-    public void leaveEvent(String par_username, String title){
-        utilDeleteParUpcomingevent(par_username,title);
-    }
 
 
-    public boolean checkIfUsernameExist(String username){
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+    public boolean checkIfEventnameExist(String eventname){
         Statement stmt = null;
         Connection conn = null;
         ResultSet rs = null;
@@ -708,7 +565,7 @@ public class ParFileUser {
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select exists(select * from parfile where username = '" + username + "');";
+            String sql = "select exists(select * from eventfile where title = '" + eventname + "');";
             stmt = conn.createStatement();
             rs = stmt.executeQuery(sql);
             rs.next();
@@ -747,33 +604,28 @@ public class ParFileUser {
         return WhetherExist;
     }
 
-    public void createPar(String username, String password){
-        utilStorePar(username, password);
-    }
-
-    public void deletePar(String username){
-        //First delete relationships with events
-        //Then delete relationships with organizers
-        //Then delete itself
-
-        //First breaking relationship with past events
-        ArrayList<String> All_past_events = utilGetAllPastEvent(username);
-        for (int i=0;i<All_past_events.size();i++){
-            utilDeleteParPastevent(username,All_past_events.get(i));
+    public void deleteEvent(String event_title){
+        //First break relationships of the event with the organizer
+        //Then break relationships of the event with all the participants
+        //Then delete the event itself
+        OrgDsFileUser temp_orgfileuser = new OrgDsFileUser();
+        ParFileUser temp_parfileuser = new ParFileUser();
+        ArrayList<String> All_past_participants = utilGetAllPastEventParticipant(event_title);
+        for (int i = 0; i < All_past_participants.size(); i ++){
+            temp_parfileuser.utilDeleteParPastevent(All_past_participants.get(i),event_title);
         }
-        ArrayList<String> All_upcoming_events = utilGetAllUpcomingEvent(username);
-        for (int i=0;i<All_upcoming_events.size();i++){
-            utilDeleteParUpcomingevent(username, All_upcoming_events.get(i));
+        ArrayList<String> All_upcoming_participants = utilGetAllUpcomingEventParticipant(event_title);
+        for (int i = 0; i < All_upcoming_participants.size(); i ++){
+            temp_parfileuser.utilDeleteParUpcomingevent(All_upcoming_participants.get(i),event_title);
         }
 
-        //Then break relationship with followings
-        ArrayList<String> All_following = utilGetAllFollowing(username);
-        for (int i=0;i<All_following.size();i++){
-            utilDeleteParFollowOrg(username, All_following.get(i));
-        }
 
-        utilDeletePar(username);
+        String organizer = utilGetOrganizer(event_title);
+        temp_orgfileuser.utilDeleteOrgPastEvent(organizer,event_title);
+        temp_orgfileuser.utilDeleteOrgUnpublishedEvent(organizer,event_title);
+        temp_orgfileuser.utilDeleteOrgUpcomingevent(organizer,event_title);
+
+        utilDeleteEvent(event_title);
 
     }
-
 }
