+++ src/main/java/screens/LoginScreen.java
@@ -6,8 +6,7 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
 public class LoginScreen extends JFrame implements ActionListener {
-    JRadioButton isParticipant = new JRadioButton();
-    JRadioButton isOrganization = new JRadioButton();
+    ButtonGroup userType = new ButtonGroup();
     JTextField username = new JTextField(15);
     /**
      * The password
@@ -21,8 +20,11 @@ public class LoginScreen extends JFrame implements ActionListener {
         JLabel title = new JLabel("Login Screen");
         title.setAlignmentX(Component.CENTER_ALIGNMENT);
 
-        LabelTextPanel userTypeInfo = new LabelTextPanel(
-                new JLabel("Choose user type"), type);
+        JRadioButton parButton = new JRadioButton("Participant");
+        parButton.setActionCommand("Participant");
+        JRadioButton orgButton = new JRadioButton("Organization");
+        orgButton.setActionCommand("Organization");
+
         LabelTextPanel usernameInfo = new LabelTextPanel(
                 new JLabel("Username"), username);
         LabelTextPanel passwordInfo = new LabelTextPanel(
@@ -30,6 +32,17 @@ public class LoginScreen extends JFrame implements ActionListener {
 
         JButton logIn = new JButton("Log in");
         JButton cancel = new JButton("Cancel");
+
+        ButtonGroup userType = new ButtonGroup();
+        userType.add(parButton);
+        userType.add(orgButton);
+
+        JPanel typeInfo = new JPanel();
+        typeInfo.add(parButton);
+        typeInfo.add(orgButton);
+
+        parButton.addActionListener(this);
+        orgButton.addActionListener(this);
 
         JPanel buttons = new JPanel();
         buttons.add(logIn);
@@ -42,6 +55,7 @@ public class LoginScreen extends JFrame implements ActionListener {
         main.setLayout(new BoxLayout(main, BoxLayout.Y_AXIS));
 
         main.add(title);
+        main.add(typeInfo);
         main.add(usernameInfo);
         main.add(passwordInfo);
         main.add(buttons);
@@ -51,12 +65,11 @@ public class LoginScreen extends JFrame implements ActionListener {
     }
     public void actionPerformed(ActionEvent evt) {
         try {
-            userLoginController.create(
-                    isParticipant.getBoolean(),
-                    isOrganization.getBoolean(),
+            userLoginController.login(
+                    userType.getActionCommand(),
                     username.getText(),
                     String.valueOf(password.getPassword()));
-            JOptionPane.showMessageDialog(this, "%s created.".formatted(username.getText()));
+            JOptionPane.showMessageDialog(this, username.getText()+" created.");
         } catch (Exception e) {
             JOptionPane.showMessageDialog(this, e.getMessage());
         }
