+++ src/test/java/tutorial/HelloWorldTest.java
@@ -1,5 +1,6 @@
 package tutorial;
 
+import HelloWorld;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
